refactor: distance, extraFare 도메인으로 검증 로직 이전
