+++ README.md
@@ -10,7 +10,7 @@
 
 ## 1단계 리팩토링 목록
 - [x] 예상치 못한 예외 발생시, 서버 관리자도 알도록 로그 추가
-- [ ] distance, extraFare 도메인에 검증 로직 작성
+- [x] distance, extraFare 도메인에 검증 로직 작성
 - [ ] 경로 검색 예외 처리 기능 추가
   - [ ] 출발지와 도착지가 구간으로 연결되어 있지 않아 경로를 찾을 수 없는 경우 예외 발생
 - [ ] calculateShortestPath, calculateShortestDistance 내부에서 최단 경로 구하는 로직이 중복 제거
