+++ src/main/java/wooteco/subway/dto/LineRequest.java
@@ -1,7 +1,5 @@
 package wooteco.subway.dto;
 
-import wooteco.subway.exception.PositiveDigitException;
-
 public class LineRequest {
 
     private String name;
@@ -15,30 +13,12 @@ public class LineRequest {
     }
 
     public LineRequest(String name, String color, Long upStationId, Long downStationId, int distance, int extraFare) {
-        validatePositive(distance, extraFare);
         this.name = name;
         this.color = color;
         this.upStationId = upStationId;
         this.downStationId = downStationId;
         this.distance = distance;
         this.extraFare = extraFare;
-    }
-
-    private void validatePositive(final int distance, final int extraFare) {
-        validatePositiveDistance(distance);
-        validatePositiveExtraFare(extraFare);
-    }
-
-    private void validatePositiveDistance(final int distance) {
-        if (distance <= 0) {
-            throw new PositiveDigitException("구간의 길이가 양수가 아닙니다.");
-        }
-    }
-
-    private void validatePositiveExtraFare(final int extraFare) {
-        if (extraFare <= 0) {
-            throw new PositiveDigitException("추가 요금이 양수가 아닙니다.");
-        }
     }
 
     public String getName() {
